refactor: move rendering logic to context package
