+++ pkg/gateway/context/middleware.go
@@ -1,4 +1,4 @@
-package middleware
+package context
 
 import (
 	"github.com/labstack/echo/v4"
