+++ pkg/gateway/context/sessions.go
@@ -1,4 +1,4 @@
-package middleware
+package context
 
 import (
 	gocontext "context"
