+++ pkg/gateway/context/vaults.go
@@ -1,4 +1,4 @@
-package middleware
+package context
 
 import (
 	"fmt"
