+++ pkg/gateway/handlers/check_handler.go
@@ -4,21 +4,21 @@ import (
 	"github.com/labstack/echo/v4"
 
 	"github.com/onsonr/sonr/internal/nebula/input"
-	"github.com/onsonr/sonr/pkg/gateway/middleware"
+	"github.com/onsonr/sonr/pkg/gateway/context"
 )
 
 // CheckProfileHandle finds the chosen handle and verifies it is unique
 func CheckProfileHandle(c echo.Context) error {
 	handle := c.FormValue("handle")
 	if handle == "" {
-		return middleware.Render(c, input.HandleError(handle, "Please enter a valid handle"))
+		return context.Render(c, input.HandleError(handle, "Please enter a valid handle"))
 	}
 	//
 	// if ok {
 	// 	return middleware.Render(c, input.HandleError(handle))
 	// }
 	//
-	return middleware.Render(c, input.HandleSuccess(handle))
+	return context.Render(c, input.HandleSuccess(handle))
 }
 
 // ValidateProfileHandle finds the chosen handle and verifies it is unique
@@ -32,5 +32,5 @@ func CheckIsHumanSum(c echo.Context) error {
 	// if intValue != data.Sum() {
 	// 	return middleware.Render(c, input.HumanSliderError(data.FirstNumber, data.LastNumber))
 	// }
-	return middleware.Render(c, input.HumanSliderSuccess())
+	return context.Render(c, input.HumanSliderSuccess())
 }
