+++ pkg/gateway/handlers/register_handler.go
@@ -2,23 +2,23 @@ package handlers
 
 import (
 	"github.com/labstack/echo/v4"
-	"github.com/onsonr/sonr/pkg/gateway/middleware"
+	"github.com/onsonr/sonr/pkg/gateway/context"
 	"github.com/onsonr/sonr/pkg/gateway/views"
 )
 
 func RenderProfileCreate(c echo.Context) error {
 	// numF, numL := middleware.GetHumanVerificationNumbers(c)
-	params := middleware.CreateProfileParams{
-		FirstNumber: int(middleware.StatusBlock(c)),
-		LastNumber:  int(middleware.StatusBlock(c)),
+	params := context.CreateProfileParams{
+		FirstNumber: int(context.StatusBlock(c)),
+		LastNumber:  int(context.StatusBlock(c)),
 	}
-	return middleware.Render(c, views.RegisterProfileView(params.FirstNumber, params.LastNumber))
+	return context.Render(c, views.RegisterProfileView(params.FirstNumber, params.LastNumber))
 }
 
 func RenderPasskeyCreate(c echo.Context) error {
-	return middleware.Render(c, views.RegisterPasskeyView("", "", "", "", ""))
+	return context.Render(c, views.RegisterPasskeyView("", "", "", "", ""))
 }
 
 func RenderVaultLoading(c echo.Context) error {
-	return middleware.Render(c, views.LoadingView())
+	return context.Render(c, views.LoadingView())
 }
